Add functions to main.py
Added the get_book_text function to read from a file and turn the book into a string.
Added the count_words function to take the string from the book and return the amount of words within in.
Added the main function that calls both get_book_text and count_words function to print the amount of words in the book on the screen.